pyproject.toml
[tool.poetry]
name = "tylkozapytam"
version = "0.1.0"
description = ""
authors = ["MarcGroc <marcingrochola1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
pre-commit = "^2.21.0"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
django-loguru = "^1.0.3"
flake8 = "^6.0.0"
pytest = "^7.2.0"
drf-spectacular = "^0.25.1"
pillow = "^9.4.0"
factory-boy = "^3.2.1"
black = "^22.12.0"
isort = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

